no event text, fix event/report links, no category for reports
