+++ lib/getPageSections.js
@@ -1,6 +1,8 @@
 const qs = require('qs');
 
 async function getPageSections(id, type = `pages`) {
+	if (!id) return;
+
 	const query = qs.stringify(
 		{
 			populate: {
